add cache for rate
