+++ package.json
@@ -19,7 +19,7 @@
   "homepage": "https://github.com/winout96/AliUAH#readme",
   "description": "",
   "devDependencies": {
-    "@types/chrome": "^0.0.236",
+    "@types/chrome": "^0.0.237",
     "husky": "^8.0.3",
     "lint-staged": "^13.2.2",
     "prettier": "2.8.8"
